# specs/001-course-assistant-mvp/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DOU-Synapse API",
    "description": "Ders materyaline bağlı, kaynaklı cevap üreten ders asistanının HTTP yüzeyi. Her cevap dayandığı sayfayla döner; dayanak yoksa uç cevap değil gerekçeli ret döndürür. Ders izolasyonu iki katmanda zorlanır: sunucu tarafında üyelik doğrulaması ve aynı oturumda PostgreSQL satır düzeyi güvenlik.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/courses/{course_id}/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "description": "Ders materyali yükler ve işleme kuyruğuna alır.\n\nDosya diske yazılmadan önce uzantı, boyut ve sihirli baytlarıyla doğrulanır; depolama\nanahtarı sunucuda üretilir, kullanıcının verdiği ad hiçbir zaman yol olarak\nkullanılmaz.\n\n**`replaces_document_id` AÇIK bir eylemdir, tahmin değil (FR-118).** Dosya adına\nbakarak otomatik eşleme reddedildi: `file_name` üzerinde hiçbir tekillik yok,\n\"hafta3.pdf\" her dönem yeniden yüklenir (yanlış pozitif) ve yeniden adlandırılmış\nbir güncelleme yakalanmaz (yanlış negatif). Yanlış işaretlenen bir soru,\nöğretmenin işarete olan güvenini bitirir; güvenilmez işaret, hiç işaret\nolmamasından kötüdür.\n\nEski belge SİLİNMEZ, yalnız `superseded_at` damgası alır: ona dayanan soruların\nkaynağı yerinde kalmalı, yoksa \"bayat\" işareti de okunamaz hâle gelirdi.",
        "operationId": "upload_document_courses__course_id__documents_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_courses__course_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_courses__course_id__documents_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut_DocumentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_courses__course_id__documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Document",
        "description": "Belgeyi ve ona bağlı tüm chunk'ları siler.\n\nChunk'lar FK cascade ile gider: materyal kaldırıldığında ondan üretilmiş vektörlerin\naramada kalmaması, \"yalnızca yüklenen materyalden cevap\" garantisinin parçasıdır.\n\nAma cascade her yere ulaşmaz: `questions.source_chunk_id` `ON DELETE RESTRICT`\ntaşır, yani o chunk'tan üretilmiş bir soru havuzdaysa silme veritabanı düzeyinde\nreddedilir. Bu kısıt bilinçlidir — kaynağı silinmiş bir soru, kaynağına karşı\ndoğrulanamayan bir sorudur (Anayasa I). Kısıtı gevşetmek yerine hatayı çeviriyoruz.",
        "operationId": "delete_document_courses__course_id__documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}/documents/{document_id}/retry": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Retry Document",
        "description": "Kalıcı hataya düşen ingestion işini eğitmen açıkça yeniden kuyruğa alır.",
        "operationId": "retry_document_courses__course_id__documents__document_id__retry_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}/documents/{document_id}/chunks": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Preview Chunks",
        "description": "Eğitmenin, belgeden ne çıkarıldığını görmesi için ilk chunk'lar.",
        "operationId": "preview_chunks_courses__course_id__documents__document_id__chunks_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChunkPreview"
                  },
                  "title": "Response Preview Chunks Courses  Course Id  Documents  Document Id  Chunks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_courses__course_id__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "PDF, PPTX, Markdown, metin veya kod dosyası"
          },
          "replaces_document_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaces Document Id",
            "description": "Bu dosyanın yerine geçtiği eski belgenin kimliği (FR-118)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_courses__course_id__documents_post"
      },
...
      "ChunkPreview": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "slide_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slide Number"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "token_count": {
            "type": "integer",
            "title": "Token Count"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "chunk_index",
          "page_number",
          "slide_number",
          "section_title",
          "content_type",
          "token_count",
          "text"
        ],
        "title": "ChunkPreview",
        "description": "Eğitmenin, sistemin belgeden ne çıkardığını görebilmesi için."
      },
...
      "DocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "supersedes_document_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Document Id"
          },
          "superseded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_name",
          "file_type",
          "byte_size",
          "status",
          "page_count",
          "chunk_count",
          "error_message",
          "created_at"
        ],
        "title": "DocumentOut"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "uploaded",
          "processing",
          "completed",
          "failed"
        ],
        "title": "DocumentStatus"
      },
      "DocumentUploadOut": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentOut"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Belge alındı ve işleme kuyruğuna eklendi."
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "DocumentUploadOut",
        "description": "Yükleme kabul yanıtı.\n\nBelge arka planda işlendiği için 202 döner; istemci `status` alanını izleyerek\nilerlemeyi gösterir."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageOut_DocumentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PageOut[DocumentOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}